Move CourseForm into own component file * link CoursePromo form w/ the course state
